# openapi (3).yaml
openapi: 3.0.0
info:
  title: 10Web API
  description: |    
    The 10Web API provides a comprehensive set of endpoints for managing websites, DNS zones, domains, SSL certificates, backups, and more. This RESTful API allows you to integrate 10Web's powerful hosting and website management capabilities into your own applications.
...
servers:
  - url: https://api.10web.io
paths:
...
  /hosting/domains/{domain_id}/staging/enable:
    post:
      summary: Enable staging for a domain
      tags:
        - Domain
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - $ref: '#/components/parameters/Authorization'
      responses:
        '200':
          description: Staging enabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /hosting/domains/{staging_domain_id}/staging/disable:
    post:
      summary: Disable staging for a domain
      tags:
        - Domain
      parameters:
        - $ref: '#/components/parameters/StagingDomainId'
        - $ref: '#/components/parameters/Authorization'
      responses:
        '200':
          description: Staging disabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /hosting/domains/{domain_id}/push-to-staging:
    post:
      summary: Push changes to staging environment
      tags:
        - Domain
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - $ref: '#/components/parameters/Authorization'
      responses:
        '200':
          description: Changes pushed to staging successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /hosting/domains/{domain_id}/backup/run:
    post:
      summary: Run a backup for a domain
      tags:
        - Backup
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - $ref: '#/components/parameters/Authorization'
      responses:
        '200':
          description: Backup started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /hosting/domains/{domain_id}/backup/{backup_id}/restore:
    post:
      summary: Restore a backup
      tags:
        - Backup
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - $ref: '#/components/parameters/BackupId'
        - $ref: '#/components/parameters/Authorization'
      responses:
        '200':
          description: Backup restored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /hosting/domains/{domain_id}/cache/enable:
    put:
      summary: Enable FastCGI caching for a domain
      tags:
        - Cache
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - $ref: '#/components/parameters/Authorization'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                cache_time:
                  type: integer
                  description: Cache TTL in minutes (must be between 1 and 1500)
                  minimum: 1
                  maximum: 1500
                  example: 600
      responses:
        '200':
          description: Cache enabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /hosting/domains/{domain_id}/cache/disable:
    put:
      summary: Disable FastCGI caching for a domain
      tags:
        - Cache
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - $ref: '#/components/parameters/Authorization'
      responses:
        '200':
          description: Cache disabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /hosting/domains/{domain_id}/cache:
    delete:
      summary: Purge all cache for a domain
      tags:
        - Cache
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - $ref: '#/components/parameters/Authorization'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                recache:
                  type: integer
                  description: Whether to rebuild the cache (1) or just clear it (0)
                  enum: [0, 1]
      responses:
        '200':
          description: Cache purged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /hosting/domains/{domain_id}/object-cache/toggle:
    post:
      summary: Toggle object cache for a domain
      tags:
        - Cache
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - $ref: '#/components/parameters/Authorization'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  description: Action to perform on object cache
                  enum: ["enable", "disable"]
                  example: "enable"
      responses:
        '200':
          description: Object cache setting updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /hosting/domains/{domain_id}/object-cache/flush:
    post:
      summary: Flush object cache for a domain
      tags:
        - Cache
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - $ref: '#/components/parameters/Authorization'
      responses:
        '200':
          description: Object cache flushed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /hosting/domains/{domain_id}/php-version/switch:
    post:
      summary: Switch PHP version for a domain
      tags:
        - PHP Management
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - $ref: '#/components/parameters/Authorization'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - php_version
              properties:
                php_version:
                  type: string
                  description: Target PHP version to switch to, valid versions need to be checked in the response of the /hosting/supported-php-versions endpoint
                  example: "8.2"
      responses:
        '200':
          description: PHP version switched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /hosting/domains/{domain_id}/php/restart:
    post:
      summary: Restart PHP for a domain
      tags:
        - PHP Management
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - $ref: '#/components/parameters/Authorization'
      responses:
        '200':
          description: PHP restarted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    DomainId:
      name: domain_id
      in: path
      required: true
      schema:
        type: integer
    StagingDomainId:
      name: staging_domain_id
      in: path
      required: true
      schema:
        type: integer
    BackupId:
      name: backup_id
      in: path
      required: true
      schema:
        type: integer
...
    Authorization:
      name: x-api-key
      in: header
      required: true
      schema:
        type: string
      description: API key for authentication
      example: "API_KEY"
  schemas:
...
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok, error]
      example:
        status: "ok"
...
  responses:
    Unauthorized:
      description: Unauthorized access
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                enum: [error]
              message:
                type: string
                example: "Unauthorized access"
    InternalServerError:
      description: Something went wrong
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                enum: [error]
              message:
                type: string
                example: "Something went wrong"
    TooManyRequests:
      description: Too many requests
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                enum: [error]
              message:
                type: string
                example: "Too many requests. Please try again later."
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                enum: [error]
              message:
                type: string
                example: "Invalid request parameters"